FRONT | COMPONENT | feat: create register form and call in his page
